just scheduler views
