+++ expo/scheduler/views/__init__.py
@@ -1,0 +1,27 @@
+from remains import (
+    get_events_display_info,
+    get_event_display_info,
+    event_list,
+    detail_view,
+    schedule_acts,
+    delete_schedule,
+    delete_event,
+    get_manage_opportunity_forms,
+    get_worker_allocation_forms,
+    allocate_workers,
+    manage_volunteer_opportunities,
+    contact_info,
+    contact_performers,
+    contact_volunteers,
+    contact_teachers,
+    contact_vendors,
+    contact_by_role,
+    set_single_role,
+    set_multi_role,
+    edit_event,
+    edit_event_display,
+    view_list,
+    calendar_view,
+)
+
+from scheduler.views.add_event import add_event
